move search & settings stores inside apps #377
