+++ app/js/App.js
@@ -3,7 +3,7 @@ import React, { Component, PropTypes } from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import { AccountActions } from './store/account'
-import { SettingsActions } from './store/settings'
+import { SettingsActions } from './account/store/settings'
 import WelcomeModal from './components/WelcomeModal'
 import { getCoreAPIPasswordFromURL, getLogServerPortFromURL } from './utils/api-utils'
 import log4js from 'log4js'
