+++ app/js/account/CreateAccountPage.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux'
 import Alert from '../components/Alert'
 import InputGroup from '../components/InputGroup'
 import { AccountActions } from '../store/account'
-import { SettingsActions } from '../store/settings'
+import { SettingsActions } from './store/settings'
 import { isPasswordValid } from '../utils'
 
 function mapStateToProps(state) {
