+++ app/js/account/store/settings/index.js
