+++ app/js/components/WelcomeModal.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux'
 import Alert from './Alert'
 import InputGroup from './InputGroup'
 import { AccountActions } from '../store/account'
-import { SettingsActions } from '../store/settings'
+import { SettingsActions } from '../account/store/settings'
 import { DROPBOX_APP_ID } from '../storage/utils/dropbox'
 import { isBackupPhraseValid } from '../utils'
