+++ app/js/profiles/store/search/actions.js
@@ -1,8 +1,8 @@
 import * as types from './types'
-import { compareProfilesByVerifications } from '../../utils/index'
+import { compareProfilesByVerifications } from '../../../utils/index'
 import log4js from 'log4js'
 
-const logger = log4js.getLogger('store/search/actions.js')
+const logger = log4js.getLogger('profiles/store/search/actions.js')
 
 function updateQuery(query) {
   return {
