+++ app/js/store/reducers/index.js
@@ -2,9 +2,9 @@ import { combineReducers } from 'redux'
 
 import { AccountReducer }  from '../account'
 import { AuthReducer }     from '../auth'
-import { SettingsReducer } from '../settings'
+import { SettingsReducer } from '../../account/store/settings'
 import { IdentityReducer } from '../identities'
-import { SearchReducer }   from '../search/index'
+import { SearchReducer }   from '../../profiles/store/search'
 
 const RootReducer = combineReducers({
   account: AccountReducer,
