+++ tests/store/search/actions/sync.test.js
@@ -1,4 +1,4 @@
-import { SearchActions } from '../../../../app/js/store/search'
+import { SearchActions } from '../../../../app/js/profiles/store/search'
 
 describe('Search Store: Sync Actions', () => {
   describe('updateQuery', () => {
