+++ tests/store/settings/actions/async.test.js
@@ -1,7 +1,7 @@
 import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
 import nock from 'nock'
-import { SettingsActions } from '../../../../app/js/store/settings'
+import { SettingsActions } from '../../../../app/js/account/store/settings'
 import btcPrice from '../../../fixtures/btcprice'
 
 const middlewares = [thunk]
