+++ tests/store/settings/reducer.test.js
@@ -1,5 +1,5 @@
-import { SettingsActions, SettingsReducer } from '../../../app/js/store/settings'
-import DEFAULT_API from '../../../app/js/store/settings/default'
+import { SettingsActions, SettingsReducer } from '../../../app/js/account/store/settings'
+import DEFAULT_API from '../../../app/js/account/store/settings/default'
 
 
 describe('Settings Store: SettingsReducer', () => {
